Added  menuitem to restaurant and some nuget updates

+++ DeliveryApp.Domain/Models/Order.cs
@@ -6,7 +6,7 @@ public class Order
 {
     public Guid Id { get; set; }
     public string ReceivedTime { get; set; }
-    public float FinalPrice { get; set; }
+    public double FinalPrice { get; set; }
     public string Status { get; set; }
     public UserDto User { get; set; }
     public Restaurant Restaurant { get; set; }
