+++ DeliveryApp.Repository/Entities/Orders.cs
@@ -4,7 +4,7 @@ public class Orders
 {
     public Guid Id { get; set; }
     public DateTime ReciviedTime { get; set; }
-    public float FinalPrice { get; set; }
+    public double FinalPrice { get; set; }
     public string Status { get; set; }
     public Users User { get; set; }
     public Restaurants Restaurant { get; set; }
